ColorFragment, ShapesFragment, NumbersFragment, AlphabetFragment all createed along with their navigation destinaton after respective buttons are clicked on landing page.
