+++ app/src/main/java/org/ohmstheresistance/milabila/fragments/AlphabetFragment.kt
@@ -5,56 +5,61 @@ import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.databinding.DataBindingUtil
+import androidx.recyclerview.widget.GridLayoutManager
 import org.ohmstheresistance.milabila.R
+import org.ohmstheresistance.milabila.databinding.AlphabetFragmentBinding
+import org.ohmstheresistance.milabila.databinding.LandingPageFragmentBinding
+import org.ohmstheresistance.milabila.dataclasses.AlphabetData
+import org.ohmstheresistance.milabila.recyclerview.AlphabetAdapter
 
-// TODO: Rename parameter arguments, choose names that match
-// the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
-private const val ARG_PARAM1 = "param1"
-private const val ARG_PARAM2 = "param2"
-
-/**
- * A simple [Fragment] subclass.
- * Use the [AlphabetFragment.newInstance] factory method to
- * create an instance of this fragment.
- */
 class AlphabetFragment : Fragment() {
-    // TODO: Rename and change types of parameters
-    private var param1: String? = null
-    private var param2: String? = null
 
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        arguments?.let {
-            param1 = it.getString(ARG_PARAM1)
-            param2 = it.getString(ARG_PARAM2)
-        }
-    }
+    private val letterList = ArrayList<AlphabetData>()
+    private val alphabetAdapter = AlphabetAdapter(letterList)
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.alphabet_fragment, container, false)
+        val alphabetFragmentBinding = DataBindingUtil.inflate<AlphabetFragmentBinding>(inflater, R.layout.alphabet_fragment, container, false)
+
+        alphabetFragmentBinding.alphabetListRecycler.layoutManager = GridLayoutManager(context, 4)
+        alphabetFragmentBinding.alphabetListRecycler.adapter = alphabetAdapter
+
+        addLetters()
+
+        return alphabetFragmentBinding.root
     }
 
-    companion object {
-        /**
-         * Use this factory method to create a new instance of
-         * this fragment using the provided parameters.
-         *
-         * @param param1 Parameter 1.
-         * @param param2 Parameter 2.
-         * @return A new instance of fragment AlphabetFragment.
-         */
-        // TODO: Rename and change types and number of parameters
-        @JvmStatic
-        fun newInstance(param1: String, param2: String) =
-            AlphabetFragment().apply {
-                arguments = Bundle().apply {
-                    putString(ARG_PARAM1, param1)
-                    putString(ARG_PARAM2, param2)
-                }
-            }
+    private fun addLetters(){
+
+        letterList.add(AlphabetData(R.drawable.a))
+        letterList.add(AlphabetData(R.drawable.b))
+        letterList.add(AlphabetData(R.drawable.c))
+        letterList.add(AlphabetData(R.drawable.d))
+        letterList.add(AlphabetData(R.drawable.e))
+        letterList.add(AlphabetData(R.drawable.f))
+        letterList.add(AlphabetData(R.drawable.g))
+        letterList.add(AlphabetData(R.drawable.h))
+        letterList.add(AlphabetData(R.drawable.i))
+        letterList.add(AlphabetData(R.drawable.j))
+        letterList.add(AlphabetData(R.drawable.k))
+        letterList.add(AlphabetData(R.drawable.l))
+        letterList.add(AlphabetData(R.drawable.m))
+        letterList.add(AlphabetData(R.drawable.n))
+        letterList.add(AlphabetData(R.drawable.o))
+        letterList.add(AlphabetData(R.drawable.p))
+        letterList.add(AlphabetData(R.drawable.q))
+        letterList.add(AlphabetData(R.drawable.r))
+        letterList.add(AlphabetData(R.drawable.s))
+        letterList.add(AlphabetData(R.drawable.t))
+        letterList.add(AlphabetData(R.drawable.u))
+        letterList.add(AlphabetData(R.drawable.v))
+        letterList.add(AlphabetData(R.drawable.w))
+        letterList.add(AlphabetData(R.drawable.x))
+        letterList.add(AlphabetData(R.drawable.y))
+        letterList.add(AlphabetData(R.drawable.z))
+
     }
-}
+   }
