+++ app/src/main/res/layout/landing_page_fragment.xml
@@ -28,12 +28,12 @@
             app:layout_constraintLeft_toLeftOf="parent"
             app:layout_constraintRight_toRightOf="parent"
             app:layout_constraintTop_toBottomOf="@id/landing_page_header_textview"
-            app:layout_constraintBottom_toTopOf="@id/landing_page_buttons_linear"
+            app:layout_constraintBottom_toTopOf="@id/landing_page_buttons_constraint"
             />
 
         <androidx.constraintlayout.widget.ConstraintLayout
 
-            android:id="@+id/landing_page_buttons_linear"
+            android:id="@+id/landing_page_buttons_constraint"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:background="@color/text_color"
